Move the Check & Synth judgements into Facet.Elab.Term.
